Give each component its own JSX file
