+++ app/pages/lab/collaborators.jsx
@@ -1,13 +1,13 @@
-import { useEffect, useRef, useState } from 'react'
-import apiClient from 'panoptes-client/lib/api-client'
-import talkClient from 'panoptes-client/lib/talk-client'
+import { useEffect, useRef, useState } from 'react';
+import apiClient from 'panoptes-client/lib/api-client';
+import talkClient from 'panoptes-client/lib/talk-client';
 
-import UserSearch from '../../components/user-search.cjsx'
-import projectSection from '../../talk/lib/project-section.coffee'
-import isAdmin from '../../lib/is-admin.coffee'
-import getAllLinked from '../../lib/get-all-linked.js'
-
-const ID_PREFIX = 'LAB_COLLABORATORS_PAGE_';
+import CollaboratorCreator from './collaborators/CollaboratorCreator.jsx';
+import UserRow from './collaborators/UserRow.jsx';
+import { ROLES_NOT_IN_TALK_API } from './collaborators/constants.js';
+import projectSection from '../../talk/lib/project-section.coffee';
+import isAdmin from '../../lib/is-admin.coffee';
+import getAllLinked from '../../lib/get-all-linked.js';
 
 const POSSIBLE_ROLES = {
   collaborator: 'admin',
@@ -21,190 +21,6 @@ const POSSIBLE_ROLES = {
 NOTE: Panoptes API and Talk API keep track of user roles separately; Panoptes can accept arbitrary role values, but Talk API can't.
 The key-value pairs of POSSIBLE_ROLES maps the roles on Panoptes API (key) to the 'comparable' roles on Talk API.
 */
-
-const ROLES_INFO = {
-  collaborator: {
-    label: 'Collaborator',
-    description: 'Collaborators have full access to edit workflows and project content, including deleting some or all of the project.'
-  },
-  expert: {
-    label: 'Expert',
-    description: 'Experts can enter "gold mode" to make authoritative gold standard classifications that will be used to validate data quality.'
-  },
-  scientist: {
-    label: 'Researcher',
-    description: 'Members of the research team will be marked as researchers on "Talk"'
-  },
-  moderator: {
-    label: 'Moderator',
-    description: 'Moderators have extra privileges in the community discussion area to moderate discussions. They will also be marked as moderators on "Talk".'
-  },
-  tester: {
-    label: 'Tester',
-    description: 'Testers can view and classify on your project to give feedback while it’s still private. If given the direct url, they can also view and classify on inactive workflows; this is useful for already launched projects that are planning on building a new workflow and woud like volunteer feedback. Testers cannot access the project builder.'
-  },
-  translator: {
-    label: 'Translator',
-    description: 'Translators will have access to the translation site.'
-  },
-  museum: {
-    label: 'Museum',
-    description: 'Enables a custom interface for the project on the Zooniverse iPad app, specifically designed to be used in a museum or exhibit space.'
-  }
-};
-  
-    
-const ROLES_NOT_IN_TALK_API = [
-  'museum'
-];
-
-
-function CollaboratorCreator({
-  onAdd,
-  possibleRoles = POSSIBLE_ROLES,
-  project = null
-}) {
-  const rolesTable = useRef();
-  const userSearch = useRef();
-  const [error, setError] = useState(null);
-  const [creating, setCreating] = useState(false);
-  let style;
-
-  if (creating) {
-    style = {
-      opacity: 0.5,
-      pointerEvents: 'none'
-    };
-  }
-
-  function handleSubmit(event) {
-    event.preventDefault();
-    const checkboxes = rolesTable.current?.querySelectorAll('[name="role"]');
-    const users = userSearch.current?.value().map(option => parseInt(option.value));
-    let roles = [];
-    checkboxes.forEach(checkbox => {
-      if (checkbox.checked) {
-        roles.push(checkbox.value);
-      }
-    });
-    let talkRoles = [];
-    Object.entries(possibleRoles).forEach(([role, talkRole]) => {
-      if (roles.includes(role) && !ROLES_NOT_IN_TALK_API.includes(role)) {
-        talkRoles.push(talkRole);
-      }
-    });
-    talkRoles = talkRoles.reduce((memo, role) => {
-      if (!memo.includes(role)) {
-        memo.push(role);
-      }
-      return memo;
-    }, []);
-
-    setError(null);
-    setCreating(true);
-
-    const newRoles = users.reduce((memo, id) => {
-      const newRoleSet = apiClient.type('project_roles').create({
-        roles,
-        links: {
-          project: project.id,
-          user: id
-        }
-      });
-      const newTalkRoleSets = talkRoles.map(role => talkClient.type('roles').create({
-        name: role,
-        section: projectSection(project),
-        user_id: id
-      }));
-      return memo.concat([newRoleSet]).concat(newTalkRoleSets);
-    }, []);
-
-    Promise.all(newRoles.map(roleSet => roleSet.save()))
-      .then((savedRoles) => {
-        userSearch.current?.clear();
-        checkboxes.forEach(checkbox => checkbox.checked = false);
-        onAdd();
-      })
-      .catch(error => {
-        if (error.message.match(/not allowed to create this role/i)) {
-          error.message = 'Your account status on this project is still being setup. Please try again later.';
-        }
-        setError(error)
-      })
-      .then(() => {
-        setCreating(false)
-      });
-  }
-
-  return(
-    <div>
-      {error &&
-        <p className="form-help error">{error.toString()}</p>
-      }
-      <form style={style}>
-        <div>
-          <UserSearch ref={userSearch} />
-        </div>
-
-        <table ref={rolesTable} className="standard-table">
-          <tbody>
-            {Object.entries(possibleRoles).map(([role, label]) => (
-              <tr key={role + '-' + label}>
-                <td><input id={ID_PREFIX + role} type="checkbox" name="role" value={role} disabled={role === 'owner'}/></td>
-                <td><strong><label htmlFor={ID_PREFIX + role}>{ROLES_INFO[role].label}</label></strong></td>
-                <td>{ROLES_INFO[role].description}</td>
-              </tr>
-            ))}
-          </tbody>
-        </table>
-
-        <p>
-          <button type="submit" className="major-button" onClick={handleSubmit}>Add user role</button>
-        </p>
-      </form>
-    </div>
-  );
-}
-
-function UserRow({
-  disabled = false,
-  onDelete,
-  onToggle,
-  possibleRoles = POSSIBLE_ROLES,
-  projectRoleSet
-}) {
-  const [user, setUser] = useState(null);
-  useEffect(function loadOwner() {
-    projectRoleSet.get('owner')
-    .then(owner => setUser(owner));
-  }, [projectRoleSet])
-
-  return (
-    <p>
-      <strong>{user?.display_name}</strong>{' '}
-      <button type="button" className="secret-button" onClick={onDelete}>&times;</button>
-      <br />
-
-      <span className="columns-container inline">
-        {Object.keys(possibleRoles).map(role => {
-          const toggleThisRole = () => onToggle(role);
-          return (
-            <label key={role}>
-              <input
-              type="checkbox"
-              name={role}
-              checked={projectRoleSet.roles.includes(role)}
-              disabled={role === 'owner' || disabled}
-              onChange={toggleThisRole}
-            />{' '}
-              {ROLES_INFO[role].label}
-            </label>
-          )
-        })}
-      </span>
-    </p>
-  );
-}
 
 export default function EditProjectCollaborators({
   possibleRoles = POSSIBLE_ROLES,
@@ -243,14 +59,12 @@ export default function EditProjectCollaborators({
         return roleSet;
       });
     })
-    .then(roleSets => setProjectRoleSets(roleSets));
+    .then(setProjectRoleSets);
   }
 
   useEffect(function fetchOwner() {
     project.get('owner')
-      .then(owner => {
-        setProjectOwner(owner);
-      })
+      .then(setProjectOwner)
   }, [project]);
 
   useEffect(function onProjectChange() {
@@ -277,7 +91,7 @@ export default function EditProjectCollaborators({
     } else {
       projectRoleSet.roles.splice(index, 1);
       const filteredRoles = projectRoleSet.talk_roles.filter(talkRole => talkRole === possibleRoles[role]);
-      filteredRoles[0]?.delete()
+      filteredRoles[0]?.delete();
     }
 
     Promise.all([projectRoleSet.update('roles').save(), talkRoleAction])
